+++ src/extensions/mvs/helpers/param-definition.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2023-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Adam Midlik <midlik@gmail.com>
  */
@@ -7,29 +7,44 @@
 import { ParamDefinition as PD } from '../../../mol-util/param-definition';
 
 
-/** Similar to `PD.Numeric` but allows leaving empty field in UI (treated as `undefined`) */
-export function MaybeIntegerParamDefinition(defaultValue?: number, info?: PD.Info): PD.Base<number | undefined> {
-    return PD.Converted<number | undefined, PD.Text>(stringifyMaybeInt, parseMaybeInt, PD.Text(stringifyMaybeInt(defaultValue), info));
+/** Similar to `PD.Numeric` but allows leaving empty field in UI (treated as `undefined`), and can only have integer values */
+export function MaybeIntegerParamDefinition(info?: PD.Info & { placeholder?: string }): PD.Converted<number | null, string> {
+    const defaultValue = null; // the default must be null, otherwise real nulls would be replaced by the default
+    return PD.Converted(stringifyMaybeInt, parseMaybeInt, PD.Text(stringifyMaybeInt(defaultValue), { ...info, disableInteractiveUpdates: true, placeholder: info?.placeholder ?? 'null' }));
 }
-/** The magic with negative zero looks crazy, but it's needed if we want to be able to write negative numbers, LOL. Please help if you know a better solution. */
-function parseMaybeInt(input: string): number | undefined {
-    if (input.trim() === '-') return -0;
+function parseMaybeInt(input: string): number | null {
     const num = parseInt(input);
-    return isNaN(num) ? undefined : num;
+    return isNaN(num) ? null : num;
 }
-function stringifyMaybeInt(num: number | undefined): string {
-    if (num === undefined) return '';
-    if (Object.is(num, -0)) return '-';
+function stringifyMaybeInt(num: number | null): string {
+    if (num === null) return '';
     return num.toString();
 }
 
+
+/** Similar to `PD.Numeric` but allows leaving empty field in UI (treated as `undefined`) */
+export function MaybeFloatParamDefinition(info?: PD.Info & { placeholder?: string }): PD.Converted<number | null, string> {
+    const defaultValue = null; // the default must be null, otherwise real nulls would be replaced by the default
+    return PD.Converted(stringifyMaybeFloat, parseMaybeFloat, PD.Text(stringifyMaybeFloat(defaultValue), { ...info, disableInteractiveUpdates: true, placeholder: info?.placeholder ?? 'null' }));
+}
+function parseMaybeFloat(input: string): number | null {
+    const num = parseFloat(input);
+    return isNaN(num) ? null : num;
+}
+function stringifyMaybeFloat(num: number | null): string {
+    if (num === null) return '';
+    return num.toString();
+}
+
+
 /** Similar to `PD.Text` but leaving empty field in UI is treated as `undefined` */
-export function MaybeStringParamDefinition(defaultValue?: string, info?: PD.Info): PD.Base<string | undefined> {
-    return PD.Converted<string | undefined, PD.Text>(stringifyMaybeString, parseMaybeString, PD.Text(stringifyMaybeString(defaultValue), info));
+export function MaybeStringParamDefinition(info?: PD.Info & { placeholder?: string }): PD.Converted<string | null, string> {
+    const defaultValue = null; // the default must be null, otherwise real nulls would be replaced by the default
+    return PD.Converted(stringifyMaybeString, parseMaybeString, PD.Text(stringifyMaybeString(defaultValue), { ...info, placeholder: info?.placeholder ?? 'null' }));
 }
-function parseMaybeString(input: string): string | undefined {
-    return input === '' ? undefined : input;
+function parseMaybeString(input: string): string | null {
+    return input === '' ? null : input;
 }
-function stringifyMaybeString(str: string | undefined): string {
-    return str === undefined ? '' : str;
+function stringifyMaybeString(str: string | null): string {
+    return str === null ? '' : str;
 }